api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
